refactor 2

+++ src/main/java/ru/job4j/chat/service/RoomService.java
@@ -1,27 +1,29 @@
 package ru.job4j.chat.service;
 
-import org.postgresql.util.PSQLException;
-import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.job4j.chat.aspect.Loggable;
 import ru.job4j.chat.exception.EntityNotFoundException;
 import ru.job4j.chat.exception.RoomNameReservedException;
-import ru.job4j.chat.exception.ServiceException;
 import ru.job4j.chat.exception.ServiceValidateException;
+import ru.job4j.chat.mapper.RoomDtoMapper;
 import ru.job4j.chat.model.Room;
 import ru.job4j.chat.repository.RoomRepository;
 
-import java.lang.reflect.InvocationTargetException;
 import java.util.*;
 
 @Service
 @Loggable
 public class RoomService implements GenericService<Room> {
 
-    private final RoomRepository roomRepository;
+    private static final String NOT_FOUND_ENTITY = "room with id = %d wasn't found";
 
-    public RoomService(RoomRepository roomRepository) {
+    private final RoomRepository roomRepository;
+    private final RoomDtoMapper roomMapper;
+
+    public RoomService(RoomRepository roomRepository, RoomDtoMapper roomMapper) {
         this.roomRepository = roomRepository;
+        this.roomMapper = roomMapper;
     }
 
     @Override
@@ -32,40 +34,39 @@ public class RoomService implements GenericService<Room> {
     @Override
     public Room getById(long id) {
         return roomRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException(String.format("room with id = %d wasn't found", id)));
+                () -> new EntityNotFoundException(String.format(NOT_FOUND_ENTITY, id)));
     }
 
     public Room getByIdWithMessages(long id) {
         return roomRepository.findByIdWithMessages(id).orElseThrow(
-                () -> new EntityNotFoundException(String.format("room with id = %d wasn't found", id)));
+                () -> new EntityNotFoundException(String.format(NOT_FOUND_ENTITY, id)));
     }
 
     @Override
     public Room save(Room room) {
+        if (roomRepository.findByName(room.getName()).isPresent()) {
+            throw new RoomNameReservedException(
+                    String.format("Room name '%s' already reserved please try again", room.getName()));
+        }
         if (room.getName() == null) {
             throw new ServiceValidateException("room name mustn't be null");
         }
-        try {
-            return roomRepository.save(room);
-        } catch (DataAccessException e) {
-            if ((e.getMostSpecificCause()) instanceof PSQLException) {
-                PSQLException sqlEx = (PSQLException) e.getMostSpecificCause();
-                if (sqlEx.getSQLState().equals("23505")) {
-                    throw new RoomNameReservedException(String.format("Room name already reserved please try again (%s)",
-                            sqlEx.getMessage()), sqlEx);
-                }
-            }
-            throw new ServiceException(String.format("Room %s can't be added (%s)", room, e.getMessage()), e);
-        }
+        return roomRepository.save(room);
     }
 
-    public Room partialUpdate(Room room) throws InvocationTargetException, IllegalAccessException {
-        Room currentRoom = getById(room.getId());
-        return save(setNewNotNullFieldsToModel(room, currentRoom));
+    public Room partialUpdate(Room room) {
+        Room oldRoom = getById(room.getId());
+        roomMapper.updateNewNotNullFieldsToRoom(room, oldRoom);
+        return save(oldRoom);
     }
 
     @Override
+    @Transactional
     public void delete(Room room) {
+        long id = room.getId();
+        if (roomRepository.findById(id).isEmpty()) {
+            throw new EntityNotFoundException(String.format("can't delete, %s %s", NOT_FOUND_ENTITY, id));
+        }
         roomRepository.delete(room);
     }
 }
